Restore and separate presentation tests (#46)
* Split up presentation tests

* Update changelog

---------
