+++ tests/test_iiif_dataclasses/fixtures/annotation3.json
