+++ tests/test_iiif_dataclasses/fixtures/georeference_annotation3.json
