+++ tests/test_iiif_dataclasses/fixtures/manifest2.json
